interpreter: options for interpreter infrastructure
The argument to Exp is annoying ... it's just a range but we never need
to use it. Need to either desugar to something else, or change Parser to
not use it (big overhaul) or just design interpreter but carrying it
around the whole time. None of which are great options.